Разделила на модули код
